CarSell -BackEnd/requirements.txt
altgraph==0.16.1
aniso8601==3.0.2
authy==2.2.1
Babel==2.6.0
backcall==0.1.0
bleach==3.0.2
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.2
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djrill==2.1.0
entrypoints==0.2.3
exponent-server-sdk==0.2.0
future==0.17.1
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-httplib2==0.0.3
gpiozero==1.4.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
httplib2==0.12.0
idna==2.7
iexfinance==0.3.4
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.5
macholib==1.11
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mpl-finance==0.10.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.4
oauth2client==4.1.3
opencv-python==3.4.4.19
ordered-set==3.0.2
pandas==0.23.4
pandas-datareader==0.7.0
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pefile==2018.8.8
phonenumbers==8.8.4
pickleshare==0.7.5
Pillow==5.3.0
plotly==3.4.1
prometheus-client==0.4.2
promise==2.2.1
prompt-toolkit==2.0.7
psutil==5.4.8
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
PyInstaller==3.4
PyJWT==1.6.4
PyLaTeX==1.3.0
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
PySocks==1.6.8
python-dateutil==2.7.5
python-dotenv==0.7.1
pytz==2018.5
pywin32-ctypes==0.2.0
pywinpty==0.5.4
pyzbar==0.1.7
pyzmq==17.1.2
qtconsole==4.4.2
reportlab==3.5.9
requests==2.20.0
requests-cache==0.4.13
retrying==1.3.3
rsa==4.0
Rx==1.6.1
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
svglib==0.8.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
twilio==6.22.0
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xlrd==1.1.0
